CalcOrtho + correction nbc
CalcConnexion n'est plus utilisé, on calcule l'orthogonalité pour les
voies dans CalOrtho. Ne fonctionne pas pour les arcs.

VoieSommets est modifié : on n'ajoute plus les sommets doubles (cas des
boucles)

+++ src/Graphe.h
@@ -34,7 +34,7 @@ public:
     QVector<long>* getArcsOfSommet(int ids) {return &(m_SomArcs[ids]);}
     //bool build_ArcsAzFromSommet(int ids, int buffer);
     bool getSommetsOfArcs(int ida, int* si, int* sf);
-    double getAngle(int ids, int ida1, int ida2);
+    float getAngle(int ids, int ida1, int ida2);
     bool checkAngle(int ids, int ida1, int ida2);
 
 private:
